debug version works. still need to fix tests
Former-commit-id: 3d1efef7039edf8765fa1c0ae85d3a1feb95b808 [formerly 1a3e1a8b4f55f55034be4d0c497e51613bd138a7]
Former-commit-id: 61fbd068e9d1b3635c44948c17cee125c81f5f8c

+++ src/node/n_constant.h
@@ -67,9 +67,9 @@ namespace FT{
 			              Stacks& stack)
     {
         if (otype == 'b')
-            stack.b.push(ArrayXb::Constant(X.cols(),int(b_value)));
+            stack.b.push(ArrayXb::Constant(X.cols(),b_value));
         else 	
-            stack.f.push(ArrayXf::Constant(X.cols(),d_value));
+            stack.f.push(ArrayXd::Constant(X.cols(),d_value));
     }
 #else
     void NodeConstant::evaluate(const MatrixXd& X, const VectorXd& y,
